requirements.txt
beautifulsoup4==4.11.1
google_api_python_client==2.65.0
google_auth_oauthlib==0.7.1
protobuf==4.21.9
python-dotenv==0.21.0
requests==2.28.1
